chore: Modify file directory for address page
